+++ demo/run_models.R
@@ -1,23 +1,26 @@
+
 library(DynareR)
 
-# Ensure that "example1.mod","example2.mod","agtrend.mod"
-# and "bkk.mod" live in the path.
+# This file should be the last to execute
 
+library(DynareR)
 
+# Copy the dynare files to the current working directory
 
+demo(agtrend)
+demo(bkk)
+demo(example1)
 
-run_models(c("example1","example2","agtrend","bkk"))
+lapply(c("agtrend","bkk","example1"),\(x) file.copy(paste0("DynareR/run_dynare/",x,"/",x,".mod"),"."))
 
-# You can create an absolute or relative path for the DynareR files.
-# The following execute existing mod files in "DynareR/run_model/"  folder
-# relative to the current path.
+run_models(c("agtrend","bkk","example1")) # This should be executed after running the demo or after running "agtrend.R", "bkk.R" and "example1.R" models
 
+run_models() # Run all models in Current Working Directory.
 
+# Copy the dynare files to the 'DynareR/run_dynare' directory
 
+lapply(c("agtrend","bkk","example1"),\(x) file.copy(paste0(x,".mod"),"DynareR/run_dynare"))
 
-# You can run all models that live in particular path as follows
-
-run_models(path="DynareR/run_models")
-
+run_models("DynareR/run_dynare*") # Run all models in 'DynareR/run_dynare' folder