+++ inst/doc/DynareR.R
@@ -7,8 +7,9 @@ knitr::opts_chunk$set(
   
 )
 
-## ----setup1, include=FALSE----------------------------------------------------
-#  library(DynareR)
+
+library(magrittr)
+library(DynareR)
 
 ## ----installation,eval=FALSE--------------------------------------------------
 #  install.packages("DynareR")
@@ -17,15 +18,17 @@ knitr::opts_chunk$set(
 #  
 #  devtools::install_github('sagirumati/DynareR')
 
-## ----includeIRF---------------------------------------------------------------
-#  include_IRF(model="",IRF="",path=".")
-
-## ----dyn----------------------------------------------------------------------
-#  write_dyn(model, code, path = ".")
+## ----IRF,out.extra='trim={0cm 7cm 0cm 7cm},clip',fig.cap="Another of figure generated from Dynare software"----
+#  
+#  include_IRF(model="bkk",IRF = "E_H2")
+#  
+#  # Alternatively, use the path argument
+#  
+#  include_IRF(path="bkk/bkk/graphs/bkk_IRF_E_H2.pdf")
 #  
 
 ## ----exampleDyn---------------------------------------------------------------
-#  DynareCodes='var y, c, k, a, h, b;
+#  dynareCodes='var y, c, k, a, h, b;
 #  varexo e, u;
 #  parameters beta, rho, alpha, delta, theta, psi, tau;
 #  alpha = 0.36;
@@ -64,14 +67,10 @@ knitr::opts_chunk$set(
 #  
 #  stoch_simul;'
 #  
-#  model<-"example1"
-#  code<-DynareCodes
 #  
-#  write_dyn(model,code)
-
-## ----mod----------------------------------------------------------------------
-#  write_mod(model, code, path = ".")
+#  write_dyn(code=dynareCodes, model="example1")
 #  
+#  write_dyn(code=dynareCodes,model="DynareR/write_dyn/example1")
 
 ## ----examplemod---------------------------------------------------------------
 #  DynareCodes='var y, c, k, a, h, b;
@@ -113,13 +112,10 @@ knitr::opts_chunk$set(
 #  
 #  stoch_simul;'
 #  
-#  model<-"example1"
-#  code<-DynareCodes
 #  
-#  write_mod(model,code)
-
-## ----runDynare----------------------------------------------------------------
-#  run_dynare(model,code,path)
+#  write_mod(model="example1",code=dynareCodes)
+#  
+#  write_mod(code=dynareCodes,model="DynareR/write_mod/example1")
 #  
 
 ## ----exampleRunDynare---------------------------------------------------------
@@ -162,26 +158,68 @@ knitr::opts_chunk$set(
 #  
 #  stoch_simul;'
 #  
-#  model<-"example1"
-#  code<-DynareCodes
-#  
-#  run_dynare(model,code)
-#  
-
-## ----runModels----------------------------------------------------------------
-#  run_models(model, path = ".")
+#  run_dynare(code=DynareCodes,model="example1",import_log = T)
+#  run_dynare(code=DynareCodes,model="DynareR/run_dynare/example1")
 #  
 
 ## ----exampleRunModels---------------------------------------------------------
 #  
-#  run_models(model=c("example1","example2","agtrend","bkk"))
+#  demo(agtrend)
+#  demo(bkk)
+#  demo(example1)
+#  
+#  # Provide the list of the `Dynare` files in a vector
+#  # Ensure that "agtrend.mod", "bkk.mod" and "example1.mod"
+#  # live in the current working directory
+#  
+#  # Copy the dynare files to the current working directory
+#  
+#  lapply(c("agtrend","bkk","example1"),\(x) file.copy(paste0(x,"/",x,".mod"),"."))
+#  
+#  run_models(c("agtrend","bkk","example1")) # Run the models in the vector.
+#  
 #  
 
 ## ----runAllDynare-------------------------------------------------------------
-#  run_models()
-#  
+#  run_models() # Run all models in Current Working Directory.
 
 ## ----runAllDynare1------------------------------------------------------------
-#  run_models(path = 'DynareR/run_models')
 #  
+#  # Copy the dynare files to the 'DynareR/run_dynare' directory
+#  
+#  lapply(c("agtrend","bkk","example1"),\(x) file.copy(paste0(x,".mod"),"DynareR/run_dynare"))
+#  
+#  run_models(model = 'DynareR/run_dynare*') # notice the * at the end
+#  
+
+## ----importLog----------------------------------------------------------------
+#  import_log(model="bkk")
+#  
+#  import_log(path="bkk/bkk.log")
+#  
+#  knitr::kable(dynare$bkk$autocorrelation)  %>% kableExtra::kable_styling(latex_options = c("basic","hold_position","scale_down")) %>%
+#   kableExtra::footnote(general="Some footnote with equation $\\alpha x^2+\\beta x+c=0$", general_title = "*",footnote_as_chunk=T,threeparttable=T,escape=F) %>%
+#  kableExtra::row_spec(0,bold=T)
+
+## ----dynareVersion,eval=F-----------------------------------------------------
+#  set_dynare_version("6-unstable-2022-04-03-0800-700a0e3a")
+#  
+
+## ----OctavePath---------------------------------------------------------------
+#  set_octave_path('C:/Program Files/GNU Octave/Octave-6.4.0/mingw64/bin/octave20.exe')
+
+## ----addPath,eval=FALSE-------------------------------------------------------
+#  
+#  add_path('/usr/lib/dynare/matlab')#  Default for Linux
+#  
+#  add_path('c:/dynare/5.1/matlab') # Default for Windows, but 5.1 can change if later version of
+#  # `Dynare` is installed.
+#  
+#  add_path('/usr/lib/dynare/matlab') # Default for macOS
+#  
+
+## ----demo---------------------------------------------------------------------
+#  demo(run_dynare)
+#  demo(run_models)
+#  demo(import_log)