+++ NEWS.md
@@ -1,3 +1,24 @@
+# Dynare 0.1.3
+
+What is new? 
+
+* Bug fixes: updated `dynare` configurations for all platforms
+
+* New function `import_log`, `set_dynare_version`, `set_octave_path` and `add_path`.
+
+* `set_dynare_version`, `set_octave_path` and `add_path` functions are only used if `Dynare` and `Octave` are not installed in the standard location.
+
+* `run_models` and `run_dynare` functions gained new argument `import_log=F`. Whether to import `Dynare` log file as a list of dataframes (`FALSE` by default)
+
+* `path` argument has been dropped in `write_dyn`, `write_mod`, `run_models` and `run_dynare` functions. You can now include the path in the `model` argument. For example `model='someModel'` for model in the current working directory, `model='some/path/to/the/somemodel'` for model in `some/path/to/the/` directory.
+
+* The default directory of `dynare` chunk is the current working directory.
+
+* Updated demo
+
+* Updated Vignettes
+
+* Updated R Markdown template
 
 # Dynare 0.1.2